+++ src/onyx/messaging/aeron/messenger.clj
@@ -23,7 +23,7 @@
 ;; TODO, make sure no stream-id collision issues
 (defmethod m/assign-task-resources :aeron
   [replica peer-id job-id task-id peer-site peer-sites]
-  {:aeron/peer-task-id nil})
+  {})
 
 (defmethod m/get-peer-site :aeron
   [peer-config]
@@ -33,7 +33,7 @@
 (defn flatten-publishers [publishers]
   (reduce into [] (vals publishers)))
 
-(defn transition-publishers [peer-config monitoring messenger publishers pub-infos]
+(defn transition-publishers [messenger-group monitoring messenger publishers pub-infos]
   (let [m-prev (into {} 
                      (map (juxt pub/key identity))
                      (flatten-publishers publishers))
@@ -47,14 +47,29 @@
          (keep (fn [k]
                  (let [old (m-prev k)
                        new (m-next k)]
-                   (reconcile-pub monitoring peer-config old new))))
+                   (reconcile-pub messenger-group monitoring old new))))
          (group-by (fn [^Publisher pub]
                      [(.dst-task-id pub) (.slot-id pub)])))))
+
+(defn pubs->random-selection-fn [task-id task->grouping-fn pubs]
+  (let [pbs (->> pubs
+                 (sort-by #(.slot-id ^Publisher %))
+                 (into-array Publisher))
+        len (count pbs)]
+    (if-let [group-fn (get task->grouping-fn task-id)]
+      (fn [segment]
+        (let [hsh (hash (group-fn segment))
+              idx (mod hsh len)]
+          (aget #^"[Lonyx.messaging.aeron.publisher.Publisher;" pbs idx)))
+      (fn [_]
+        (let [idx (.nextInt (java.util.concurrent.ThreadLocalRandom/current) 0 len)]
+          (aget #^"[Lonyx.messaging.aeron.publisher.Publisher;" pbs idx))))))
 
 (deftype AeronMessenger [messenger-group 
                          monitoring
                          id 
-                         ticket-counters 
+                         task->grouping-fn
+                         control-message-buf
                          ^:unsynchronized-mutable replica-version 
                          ^:unsynchronized-mutable epoch 
                          ^:unsynchronized-mutable publishers 
@@ -66,7 +81,7 @@
 
   (stop [component]
     (run! pub/stop (flatten-publishers publishers))
-    (when subscriber (sub/stop subscriber))
+    (some-> subscriber sub/stop)
     (set! replica-version nil)
     (set! epoch nil)
     (set! publishers nil)
@@ -80,30 +95,25 @@
   (publishers [messenger]
     (flatten-publishers publishers))
 
-  (task->publishers [messenger dst-task-id]
-    (get task-publishers dst-task-id))
+  (task->publishers [messenger]
+    task-publishers)
 
   (subscriber [messenger]
     subscriber)
 
   (info [messenger]
-    {:ticket-counters @ticket-counters
-     :replica-version replica-version
+    {:replica-version replica-version
      :epoch epoch
      :channel (autil/channel (:peer-config messenger-group))
      :publishers (mapv pub/info (m/publishers messenger))
      :subscriber (sub/info subscriber)})
 
   (update-publishers [messenger pub-infos]
-    (set! publishers (transition-publishers (:peer-config messenger-group) 
-                                            monitoring
-                                            messenger publishers pub-infos))
+    (set! publishers (transition-publishers messenger-group monitoring messenger publishers pub-infos))
     (set! task-publishers (->> (flatten-publishers publishers)
                                (group-by #(second (.dst-task-id ^Publisher %)))
-                               (map (fn [[k pubs]]
-                                      [k (->> pubs
-                                              (sort-by #(.slot-id ^Publisher %))
-                                              (into []))]))
+                               (map (fn [[task-id pubs]]
+                                      [task-id (pubs->random-selection-fn task-id task->grouping-fn pubs)]))
                                (into {})))
     messenger)
 
@@ -113,17 +123,10 @@
           (sub/update-sources! 
            (or subscriber
                (sub/start 
-                (new-subscription (:peer-config messenger-group) 
-                                  monitoring
-                                  id
-                                  ticket-counters
-                                  sub-info)))
+                (new-subscription messenger-group monitoring id sub-info)))
            (:sources sub-info)))
     (assert subscriber)
     messenger)
-
-  (ticket-counters [messenger]
-    ticket-counters)
 
   (set-replica-version! [messenger rv]
     (assert (or (nil? replica-version) (> rv replica-version)) [rv replica-version])
@@ -153,12 +156,12 @@
 
   (offer-barrier [messenger publisher barrier-opts]
     (let [barrier (merge (t/barrier replica-version epoch (pub/short-id publisher)) barrier-opts)
-          buf (sz/serialize barrier)]
-      (let [ret (pub/offer! publisher buf (.capacity buf) (dec epoch))] 
+          len (sz/serialize control-message-buf 0 barrier)]
+      (let [ret (pub/offer! publisher control-message-buf len (dec epoch))] 
         (debug "Offer barrier:" [:ret ret :message barrier :pub (pub/info publisher)])
         ret))))
 
-(defmethod m/build-messenger :aeron [peer-config messenger-group monitoring id]
-  (->AeronMessenger messenger-group monitoring id 
-                    (:ticket-counters messenger-group) 
+(defmethod m/build-messenger :aeron [peer-config messenger-group monitoring id task->grouping-fn]
+  (->AeronMessenger messenger-group monitoring id task->grouping-fn
+                    (UnsafeBuffer. (byte-array onyx.types/max-control-message-size))
                     nil nil nil nil nil))